Fixed the onAuthStateChangedOrReloaded to be a broadcast Stream
Fixed the onAuthStateChangedOrReloaded to be a broadcast Stream, as intended originally.
Added a await for the old user reload as this is more consistent and what's expected, even though it's not necessary, as reload is just a MethodChannel call.
Fixed a typo.
Updated the tests to check if onUserReloaded and onAuthStateChangedOrReloaded are broadcast streams.
Reformatted the tests.